editing landings
