+++ package.json
@@ -34,6 +34,7 @@
     "axios": "^0.18.0",
     "babel": "^6.23.0",
     "firebase": "^5.9.1",
+    "history": "^4.9.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.3",
     "react-dom": "^16.8.3",
