+++ src/assets/layout/home.styl
@@ -1,31 +1,17 @@
-.home
-  padding-top 50px
-  text-align center
-  table
-    margin-top 50px
-    text-align left
-    width 100%
-    th,
-    td
-      padding 10px 5px
-      vertical-align top
-    td
-      border-top 1px solid #dee2e6
-.modal-create
-  strong
-    display block
-  form
+.pre-home
+  background-color $Blue
+  height 100vh
+  .container
+    align-items center
+    display flex
+    height 100%
+    justify-content space-around
+  h3
+    color white
     display inline-block
-    margin-top 30px
-    max-width 450px
-    text-align center
-    input,
-    select
-      display block
-      margin-bottom 10px
-      width 100%
-.modal-delete
-  .alert
-    display block
-    color red
-    margin-bottom 25px
+  a
+    background-color $Lemon
+    color $Blue
+    display inline-block
+    padding 10px 20px
+    text-decoration none
