+++ src/assets/layout/landings.styl
@@ -1,0 +1,42 @@
+.landings
+  text-align center
+  .container
+    padding 50px 0
+  table
+    margin-top 50px
+    text-align left
+    width 100%
+    th,
+    td
+      padding 10px 5px
+      vertical-align top
+    td
+      border-top 1px solid #dee2e6
+.modal-create
+  .modal-header
+    background-color $Blue
+    height 50px
+    left 0
+    padding 15px 10px
+    position absolute
+    text-align center
+    top 0
+    width 100%
+    strong
+      color $Lemon
+      display block
+  form
+    display inline-block
+    margin-top 70px
+    max-width 450px
+    text-align center
+    input,
+    select
+      display block
+      margin-bottom 10px
+      width 100%
+.modal-delete
+  .alert
+    display block
+    color red
+    margin-bottom 25px
