+++ src/assets/main.styl
@@ -3,5 +3,6 @@
 @import './base.styl'
 @import './module/*'
 @import './layout/home.styl'
+@import './layout/landings.styl'
 @import './layout/create.styl'
 @import './layout/result.styl'
